Move the pdo drivers into their own directories
This is to make it more sane to add automated detection of the dependencies between extensions. Without it,
