+++ hphp/runtime/ext/pdo/config.cmake
@@ -8,21 +8,3 @@ HHVM_DEFINE_EXTENSION("pdo"
   SYSTEMLIB
     ext_pdo.php
 )
-
-HHVM_DEFINE_EXTENSION("pdo_mysql"
-  SOURCES
-    pdo_mysql.cpp
-  HEADERS
-    pdo_mysql.h
-  DEPENDS
-    libMySQL
-)
-
-HHVM_DEFINE_EXTENSION("pdo_sqlite"
-  SOURCES
-    pdo_sqlite.cpp
-  HEADERS
-    pdo_sqlite.h
-  DEPENDS
-    libSQLite
-)
