+++ hphp/runtime/ext/pdo_mysql/config.cmake
@@ -1,0 +1,8 @@
+HHVM_DEFINE_EXTENSION("pdo_mysql"
+  SOURCES
+    pdo_mysql.cpp
+  HEADERS
+    pdo_mysql.h
+  DEPENDS
+    libMySQL
+)
