+++ hphp/runtime/ext/pdo_sqlite/config.cmake
@@ -1,0 +1,8 @@
+HHVM_DEFINE_EXTENSION("pdo_sqlite"
+  SOURCES
+    pdo_sqlite.cpp
+  HEADERS
+    pdo_sqlite.h
+  DEPENDS
+    libSQLite
+)
