+++ hphp/runtime/ext/pdo_sqlite/pdo_sqlite.h
